backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
pydantic-settings==2.5.2
playwright==1.47.0
alembic==1.13.2
apscheduler==3.10.4
httpx==0.27.0
dateparser==1.2.0
argon2-cffi==23.1.0
cryptography==43.0.1
python-jose==3.3.0
email-validator==2.2.0
